fix:logo added with navigation to homepage issue:7
