+++ app/catalog/[chainId]/[catalogContractAddress]/page.tsx
@@ -9,8 +9,8 @@ import * as Alert from 'components/park-ui/alert';
 import { Heading } from 'components/park-ui/heading';
 import { Text } from 'components/park-ui/text';
 import { useMediaQuery } from 'lib/hooks/use-media-query';
-import { getIsSupportedChainId } from 'lib/utils/get-is-supported-chain-id';
-import { jsonStringifyWithBigint } from 'lib/utils/json-stringify-with-bigint';
+import { getIsSupportedChainId } from 'components/common/logo/utils/get-is-supported-chain-id';
+import { jsonStringifyWithBigint } from 'components/common/logo/utils/json-stringify-with-bigint';
 import { useReadRmrkCatalogImplGetPaginatedPartIds } from 'lib/wagmi/generated';
 import { AlertCircle } from 'lucide-react';
 import React from 'react';
